--- benchmark/env.js
@@ -1,9 +1,0 @@
-/*global suite, bench */
-'use strict';
-var generators = require('..');
-
-suite('Env', function () {
-  bench('#lookup()', function () {
-    generators().lookup();
-  });
-});
