+++ lib/actions/invoke.js
@@ -1,17 +1,5 @@
 'use strict';
 
-var GroupedQueue = require('grouped-queue');
-
-// Local Environment#duplicate copy before to work with older yo versions.
-// Delete once yo is updated and been out a while.
-var duplicateEnvironment = function (initialEnv) {
-  var queues = require('../env').queues;
-  // Hack: create a clone of the environment with a new instance of runLoop
-  var env = Object.create(initialEnv);
-  env.runLoop = new GroupedQueue(queues);
-  return env;
-};
-
 /**
  * Receives a `namespace`, and an Hash of `options` to invoke a given
  * generator. The usual list of arguments can be set with `options.args`
@@ -35,7 +23,7 @@ module.exports = function invoke(namespace, options, cb) {
 
   // Hack: create a clone of the environment because we don't want to share
   // the runLoop
-  var env = duplicateEnvironment(this.env);
+  var env = require('yeoman-environment').util.duplicateEnv(this.env);
   var generator = env.create(namespace, options);
 
   this.log.emit('up');
