+++ lib/test/helpers.js
@@ -11,6 +11,7 @@ var async = require('async');
 var rimraf = require('rimraf');
 var mkdirp = require('mkdirp');
 var runAsync = require('run-async');
+var yeoman = require('yeoman-environment');
 var assert = require('./assert');
 var generators = require('../..');
 var RunContext = require('../test/run-context');
@@ -280,7 +281,7 @@ exports.createDummyGenerator = function () {
  */
 
 exports.createGenerator = function (name, dependencies, args, options) {
-  var env = generators();
+  var env = yeoman.createEnv();
 
   this.registerDependencies(env, dependencies);
