+++ lib/test/run-context.js
@@ -2,7 +2,7 @@
 var path = require('path');
 var assert = require('assert');
 var _ = require('lodash');
-var yeoman = require('../..');
+var yeoman = require('yeoman-environment');
 var util = require('util');
 var EventEmitter = require('events').EventEmitter;
 var helpers = require('./helpers');
@@ -53,7 +53,7 @@ RunContext.prototype._run = function () {
   this.runned = true;
 
   var namespace;
-  this.env = yeoman();
+  this.env = yeoman.createEnv();
 
   helpers.registerDependencies(this.env, this._dependencies);
